backend/requirements.txt
azure-functions
openai
azure-cosmos
requests